move story cards to top level state (frame) so <nav> can search it
