+++ src/site/shared/nav/nav.html.js
@@ -1,12 +1,38 @@
-
-
-
-
+/*
+ * Title: ./src/site/shared/nav/nav.html.js
+ *
+ * Description: all HTML for the nav bar
+ *
+ * Author: Ben Merchant
+*/
 export const navTemplate = `
-<nav class="navbar navbar-expand-lg navbar-light bg-dark">
-  <a href="#" class="navbar-brand text-white">Ben Merchant dot dev</a>
-  <button type="button" name="menu" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+<nav class="navbar navbar-expand-lg navbar-dark bg-dark">
+  <a class="navbar-brand" href="#">Ben Merchant dot dev</a>
+  <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarTogglerDemo02" aria-controls="navbarTogglerDemo02" aria-expanded="false" aria-label="Toggle navigation">
     <span class="navbar-toggler-icon"></span>
   </button>
+
+  <div class="collapse navbar-collapse" id="navbarTogglerDemo02">
+    <div class="navbar-nav mr-auto mt-2 mt-lg-0">
+
+        <a class="nav-item nav-link" ui-sref="home">All</a>
+
+
+        <a class="nav-item nav-link" href="#">About</a>
+        <a class="nav-item nav-link" href="#">Tutorials</a>
+        <a class="nav-item nav-link" href="#">Projects</a>
+        <a class="nav-item nav-link" href="#">Articles</a>
+        <a class="nav-item nav-link" href="#">Blogs</a>
+        <a class="nav-item nav-link" ui-sref="linkList">Link List</a>
+
+    </div>
+    <form class="form-inline my-2 my-lg-0">
+      <input class="form-control mr-sm-2" type="search" placeholder="Search">
+      <button class="btn btn-outline-success my-2 my-sm-0" type="submit">Search</button>
+    </form>
+  </div>
 </nav>
 `;
+
+// <button class="btn btn-success" ui-sref="home">Home</button>
+// <button class="btn btn-success" ui-sref="linkList">Link List</button>
